Updated style panel
