+++ changelog.txt
@@ -2,6 +2,7 @@
 Version: 0.11.14
 Date: 2020-10-21
   Changes:
+    - Updated style panel
     - Updated help content, more information added
     - Added some control tooltip information
     - Added steam heat ingredient for machine with heat energy
