+++ core/defines.lua
@@ -72,17 +72,7 @@ helmod_frame_style = {
   panel = "helmod_frame_panel",
   cell = "helmod_frame_hidden",
   tab = "helmod_frame_tab",
-  section = "helmod_frame_section",
-  scroll_pane = "scroll_pane",
-  scroll_recipe_selector = "helmod_scroll_recipe_selector",
-  recipe_column = "helmod_frame_recipe_info"
-}
-
-helmod_scroll_style = {
-  default = "scroll_pane",
-  recipe_selector = "helmod_scroll_recipe_selector",
-  recipe_list = "helmod_scroll_recipe_module_list",
-  pin_tab = "helmod_scroll_block_pin_tab"
+  section = "helmod_frame_section"
 }
 
 helmod_table_style = {
