+++ tab/AbstractTab.lua
@@ -51,38 +51,6 @@ function AbstractTab:getPanelName()
 end
 
 -------------------------------------------------------------------------------
--- Get or create index panel
---
--- @function [parent=#AbstractTab] getIndexPanel
---
-function AbstractTab:getIndexPanel()
-  local flow_panel, content_panel, menu_panel = self:getPanel()
-  local panel_name = "index_panel"
-  if content_panel[panel_name] ~= nil and content_panel[panel_name].valid then
-    return content_panel[panel_name]["table_index"]
-  end
-  local panel = GuiElement.add(content_panel, GuiFrameH(panel_name):style(helmod_frame_style.panel))
-  panel.style.horizontally_stretchable = true
-  local table_index = GuiElement.add(panel, GuiTable("table_index"):column(GuiElement.getIndexColumnNumber()):style(helmod_table_style.list))
-  return table_index
-end
-
--------------------------------------------------------------------------------
--- Get or create model panel
---
--- @function [parent=#AbstractTab] getDebugPanel
---
-function AbstractTab:getDebugPanel()
-  local flow_panel, content_panel, menu_panel = self:getPanel()
-  local panel_name = "debug_panel"
-  if content_panel[panel_name] ~= nil and content_panel[panel_name].valid then
-    return content_panel[panel_name]
-  end
-  local panel = GuiElement.add(content_panel, GuiFrameH(panel_name):style(helmod_frame_style.panel):caption("Debug"))
-  return panel
-end
-
--------------------------------------------------------------------------------
 -- Get or create info panel
 --
 -- @function [parent=#AbstractTab] getInfoPanel
@@ -101,19 +69,19 @@ function AbstractTab:getInfoPanel()
   local info_panel = GuiElement.add(panel, GuiFlowV("info"))
   GuiElement.add(info_panel, GuiLabel("label-info"):caption({"helmod_common.information"}):style("helmod_label_title_frame"))
   GuiElement.setStyle(info_panel, "block_info", "width")
-  local info_scroll = GuiElement.add(info_panel, GuiScroll("info-scroll"):style(helmod_frame_style.scroll_pane))
+  local info_scroll = GuiElement.add(info_panel, GuiScroll("info-scroll"):style("helmod_scroll_pane"))
   info_scroll.style.horizontally_stretchable = true
 
   local output_panel = GuiElement.add(panel, GuiFlowV("output"))
   GuiElement.add(output_panel, GuiLabel("label-info"):caption({"helmod_common.output"}):style("helmod_label_title_frame"))
   GuiElement.setStyle(output_panel, "block_info", "height")
-  local output_scroll = GuiElement.add(output_panel, GuiScroll("output-scroll"):style(helmod_frame_style.scroll_pane))
+  local output_scroll = GuiElement.add(output_panel, GuiScroll("output-scroll"):style("helmod_scroll_pane"))
 
 
   local input_panel = GuiElement.add(panel, GuiFlowV("input"))
   GuiElement.add(input_panel, GuiLabel("label-info"):caption({"helmod_common.input"}):style("helmod_label_title_frame"))
   GuiElement.setStyle(input_panel, "block_info", "height")
-  local input_scroll = GuiElement.add(input_panel, GuiScroll("input-scroll"):style(helmod_frame_style.scroll_pane))
+  local input_scroll = GuiElement.add(input_panel, GuiScroll("input-scroll"):style("helmod_scroll_pane"))
   return info_scroll, output_scroll, input_scroll
 end
 
@@ -135,12 +103,11 @@ function AbstractTab:getInfoPanel2()
 
   local options_panel = GuiElement.add(panel, GuiFlowV("options"))
   GuiElement.add(options_panel, GuiLabel("label-info"):caption({"helmod_button.options"}):style("helmod_label_title_frame"))
-  --local options_scroll = GuiElement.add(options_panel, GuiScroll("options-scroll"):style(helmod_frame_style.scroll_pane))
   local options_scroll = GuiElement.add(options_panel, GuiFlowV("options-scroll"))
 
   local info_panel = GuiElement.add(panel, GuiFlowV("info"))
   GuiElement.add(info_panel, GuiLabel("label-info"):caption({"helmod_common.information"}):style("helmod_label_title_frame"))
-  local info_scroll = GuiElement.add(info_panel, GuiScroll("info-scroll"):style(helmod_frame_style.scroll_pane))
+  local info_scroll = GuiElement.add(info_panel, GuiScroll("info-scroll"):style("helmod_scroll_pane"))
 
   local left_panel = GuiElement.add(panel, GuiFlowV("left-info"))
   GuiElement.setStyle(left_panel, "block_info", "height")
@@ -187,7 +154,7 @@ function AbstractTab:getLeftInfoPanel2()
   local tool_panel = GuiElement.add(header_panel, GuiFlowH(tool_name))
   --tool_panel.style.horizontally_stretchable = true
   --tool_panel.style.horizontal_align = "right"
-  local scroll_panel = GuiElement.add(parent_panel, GuiScroll(panel_name):style(helmod_frame_style.scroll_pane))
+  local scroll_panel = GuiElement.add(parent_panel, GuiScroll(panel_name):style("helmod_scroll_pane"))
   scroll_panel.style.horizontally_stretchable = true
 
   return label_panel, tool_panel, scroll_panel
@@ -212,7 +179,7 @@ function AbstractTab:getRightInfoPanel2()
   local tool_panel = GuiElement.add(header_panel, GuiFlowH(tool_name))
   --tool_panel.style.horizontally_stretchable = true
   --tool_panel.style.horizontal_align = "right"
-  local scroll_panel = GuiElement.add(parent_panel, GuiScroll(panel_name):style(helmod_frame_style.scroll_pane))
+  local scroll_panel = GuiElement.add(parent_panel, GuiScroll(panel_name):style("helmod_scroll_pane"))
   scroll_panel.style.horizontally_stretchable = true
 
   return label_panel, tool_panel, scroll_panel
@@ -224,12 +191,10 @@ end
 -- @function [parent=#AbstractTab] getResultPanel
 --
 function AbstractTab:getResultPanel()
-  local flow_panel, content_panel, menu_panel = self:getPanel()
-  local panel_name = "result"
-  if content_panel[panel_name] ~= nil and content_panel[panel_name].valid then
-    return content_panel[panel_name]
+  local panel = self:getFramePanel("result")
+  if panel["label_panel"] == nil then
+    GuiElement.add(panel, GuiLabel("label_panel"):caption(self:getButtonCaption()):style("heading_1_label"))
   end
-  local panel = GuiElement.add(content_panel, GuiFrameV(panel_name):caption(self:getButtonCaption()))
   panel.style.horizontally_stretchable = true
   panel.style.vertically_stretchable = true
   return panel
@@ -241,17 +206,19 @@ end
 -- @function [parent=#AbstractTab] getResultPanel2
 --
 function AbstractTab:getResultPanel2()
-  local flow_panel, content_panel, menu_panel = self:getPanel()
-  local panel_name = "result2"
-  if content_panel[panel_name] ~= nil and content_panel[panel_name].valid then
-    return content_panel[panel_name]["result1"],content_panel[panel_name]["result2"]
+  local panel = self:getFlowPanel("result", "horizontal")
+  local panel_name = "table_result"
+  if panel[panel_name] ~= nil and panel[panel_name].valid then
+    return panel[panel_name]["result1"], panel[panel_name]["result2"]
   end
-  local panel = GuiElement.add(content_panel, GuiTable(panel_name):column(2):style(helmod_table_style.panel))
-  panel.style.horizontally_stretchable = true
-  panel.style.vertically_stretchable = true
 
-  local panel1 = GuiElement.add(panel, GuiFrameV("result1"))
-  local panel2 = GuiElement.add(panel, GuiFrameV("result2"):caption(self:getButtonCaption()))
+  local table_panel = GuiElement.add(panel, GuiTable(panel_name):column(2):style("helmod_table_panel"))
+  table_panel.style.horizontally_stretchable = true
+  table_panel.style.vertically_stretchable = true
+
+  local panel1 = GuiElement.add(table_panel, GuiFrameV("result1"))
+  local panel2 = GuiElement.add(table_panel, GuiFrameV("result2"))
+  GuiElement.add(panel2, GuiLabel("label_panel"):caption(self:getButtonCaption()):style("heading_1_label"))
   panel2.style.horizontally_stretchable = true
   panel2.style.vertically_stretchable = true
   return panel1,panel2
@@ -267,7 +234,7 @@ function AbstractTab:getResultScrollPanel()
   if parent_panel["scroll-data"] ~= nil and parent_panel["scroll-data"].valid then
     return parent_panel["scroll-data"]
   end
-  local scroll_panel = GuiElement.add(parent_panel, GuiScroll("scroll-data"):style(helmod_frame_style.scroll_pane))
+  local scroll_panel = GuiElement.add(parent_panel, GuiScroll("scroll-data"):style("helmod_scroll_pane"))
   scroll_panel.style.horizontally_stretchable = true
   scroll_panel.style.vertically_stretchable = true
   return scroll_panel
@@ -285,11 +252,11 @@ function AbstractTab:getResultScrollPanel2()
   end
   local header_panel1 = GuiElement.add(parent_panel1, GuiFlowV("header-data1"))
   local header_panel2 = GuiElement.add(parent_panel2, GuiFlowV("header-data2"))
-  local scroll_panel1 = GuiElement.add(parent_panel1, GuiScroll("scroll-data1"):style(helmod_frame_style.scroll_pane))
+  local scroll_panel1 = GuiElement.add(parent_panel1, GuiScroll("scroll-data1"):style("helmod_scroll_pane"))
   --scroll_panel1.style.horizontally_stretchable = true
   scroll_panel1.style.vertically_stretchable = true
   scroll_panel1.style.width = 70
-  local scroll_panel2 = GuiElement.add(parent_panel2, GuiScroll("scroll-data2"):style(helmod_frame_style.scroll_pane))
+  local scroll_panel2 = GuiElement.add(parent_panel2, GuiScroll("scroll-data2"):style("helmod_scroll_pane"))
   scroll_panel2.style.horizontally_stretchable = true
   scroll_panel2.style.vertically_stretchable = true
   return header_panel1,header_panel2,scroll_panel1,scroll_panel2
@@ -327,10 +294,11 @@ function AbstractTab:updateMenuPanel(event)
   local current_block = User.getParameter("current_block")
 
   -- action panel
-  local action_panel = self:getLeftMenuPanel()
-  action_panel.clear()
+  local left_panel, right_panel = self:getMenuPanel()
+  left_panel.clear()
+  right_panel.clear()
 
-  local group0 = GuiElement.add(action_panel, GuiFlowH("group0"))
+  local group0 = GuiElement.add(left_panel, GuiFlowH("group0"))
   for _, form in pairs(Controller.getViews()) do
     if string.find(form.classname, "Tab") and form:isVisible() and not(form:isSpecial()) then
       local icon_hovered, icon = form:getButtonSprites()
@@ -341,18 +309,18 @@ function AbstractTab:updateMenuPanel(event)
   end
 
   if self.classname == "HMAdminTab" then
-    GuiElement.add(action_panel, GuiButton("HMRuleEdition", "OPEN"):caption({"helmod_result-panel.add-button-rule"}))
-    GuiElement.add(action_panel, GuiButton(self.classname, "reset-rules"):caption({"helmod_result-panel.reset-button-rule"}))
+    GuiElement.add(left_panel, GuiButton("HMRuleEdition", "OPEN"):caption({"helmod_result-panel.add-button-rule"}))
+    GuiElement.add(left_panel, GuiButton(self.classname, "reset-rules"):caption({"helmod_result-panel.reset-button-rule"}))
   else
     -- add recipe
-    local group1 = GuiElement.add(action_panel, GuiFlowH("group1"))
+    local group1 = GuiElement.add(left_panel, GuiFlowH("group1"))
     local block_id = current_block or "new"
     GuiElement.add(group1, GuiButton("HMRecipeSelector", "OPEN", block_id):sprite("menu", "wrench-white", "wrench"):style("helmod_button_menu"):tooltip({"helmod_result-panel.add-button-recipe"}))
     GuiElement.add(group1, GuiButton("HMTechnologySelector", "OPEN", block_id):sprite("menu", "graduation-white", "graduation"):style("helmod_button_menu"):tooltip({"helmod_result-panel.add-button-technology"}))
     GuiElement.add(group1, GuiButton("HMEnergySelector", "OPEN", block_id):sprite("menu", "nuclear-white","nuclear"):style("helmod_button_menu"):tooltip({"helmod_result-panel.select-button-energy"}))
     GuiElement.add(group1, GuiButton("HMRecipeExplorer", "OPEN", block_id):sprite("menu", "search-white", "search"):style("helmod_button_menu"):tooltip({"helmod_button.explorer"}))
 
-    local group2 = GuiElement.add(action_panel, GuiFlowH("group2"))
+    local group2 = GuiElement.add(left_panel, GuiFlowH("group2"))
     -- copy past
     GuiElement.add(group2, GuiButton(self.classname, "copy-model", model.id):sprite("menu", "copy-white", "copy"):style("helmod_button_menu"):tooltip({"helmod_button.copy"}))
     GuiElement.add(group2, GuiButton(self.classname, "past-model", model.id):sprite("menu", "paste-white", "paste"):style("helmod_button_menu"):tooltip({"helmod_button.past"}))
@@ -373,14 +341,14 @@ function AbstractTab:updateMenuPanel(event)
     -- refresh control
     GuiElement.add(group2, GuiButton(self.classname, "refresh-model", model.id):sprite("menu", "refresh-white", "refresh"):style("helmod_button_menu"):tooltip({"helmod_result-panel.refresh-button"}))
 
-    local group3 = GuiElement.add(action_panel, GuiFlowH("group3"))
+    local group3 = GuiElement.add(left_panel, GuiFlowH("group3"))
     -- pin control
     if self.classname == "HMProductionBlockTab" then
       GuiElement.add(group3, GuiButton("HMPinPanel", "OPEN", block_id):sprite("menu", "pin-white", "pin"):style("helmod_button_menu"):tooltip({"helmod_result-panel.tab-button-pin"}))
       GuiElement.add(group3, GuiButton("HMSummaryPanel", "OPEN", block_id):sprite("menu", "brief-white","brief"):style("helmod_button_menu"):tooltip({"helmod_result-panel.tab-button-summary"}))
       -- Model Debug
       if User.getModGlobalSetting("debug_solver") == true then
-        local groupDebug = GuiElement.add(action_panel, GuiFlowH("groupDebug"))
+        local groupDebug = GuiElement.add(left_panel, GuiFlowH("groupDebug"))
         GuiElement.add(groupDebug, GuiButton("HMModelDebug", "OPEN", block_id):sprite("menu", "bug-white", "bug"):style("helmod_button_menu"):tooltip("Open Debug"))
       end
       
@@ -391,7 +359,7 @@ function AbstractTab:updateMenuPanel(event)
     end
 
     -- preferences
-    local groupPref = GuiElement.add(action_panel, GuiFlowH("groupPref"))
+    local groupPref = GuiElement.add(left_panel, GuiFlowH("groupPref"))
     GuiElement.add(groupPref, GuiButton("HMModelEdition", "OPEN", block_id):sprite("menu", "edit-white", "edit"):style("helmod_button_menu"):tooltip({"helmod_button.preferences"}))
     GuiElement.add(groupPref, GuiButton("HMPreferenceEdition", "OPEN", block_id):sprite("menu", "services-white", "services"):style("helmod_button_menu"):tooltip({"helmod_button.preferences"}))
     
@@ -403,7 +371,7 @@ function AbstractTab:updateMenuPanel(event)
     -- logistics
     if display_logistic_row == true then
       local logistic_row_item = User.getParameter("logistic_row_item") or "belt"
-      local logistic2 = GuiElement.add(action_panel, GuiFlowH("logistic2"))
+      local logistic2 = GuiElement.add(left_panel, GuiFlowH("logistic2"))
       for _,type in pairs({"inserter", "belt", "container", "transport"}) do
         local item_logistic = Player.getDefaultItemLogistic(type)
         local style = "helmod_button_menu"
@@ -413,7 +381,7 @@ function AbstractTab:updateMenuPanel(event)
       end
       
       local logistic_row_fluid = User.getParameter("logistic_row_fluid") or "pipe"
-      local logistic3 = GuiElement.add(action_panel, GuiFlowH("logistic3"))
+      local logistic3 = GuiElement.add(left_panel, GuiFlowH("logistic3"))
       for _,type in pairs({"pipe", "container", "transport"}) do
         local fluid_logistic = Player.getDefaultFluidLogistic(type)
         local style = "helmod_button_menu"
@@ -425,10 +393,7 @@ function AbstractTab:updateMenuPanel(event)
   end
 
 
-  local time_panel = self:getRightMenuPanel()
-  time_panel.clear()
-
-  local group_special = GuiElement.add(time_panel, GuiFlowH("group_special"))
+  local group_special = GuiElement.add(right_panel, GuiFlowH("group_special"))
   GuiElement.add(group_special, GuiButton("HMRichTextPanel", "OPEN"):sprite("menu", "text-white", "text"):style("helmod_button_menu"):tooltip({"helmod_panel.richtext"}))
   GuiElement.add(group_special, GuiButton("HMCalculator", "OPEN"):sprite("menu", "calculator-white", "calculator"):style("helmod_button_menu"):tooltip({"helmod_calculator-panel.title"}))
 
@@ -441,7 +406,7 @@ function AbstractTab:updateMenuPanel(event)
     end
   end
 
-  local group_time = GuiElement.add(time_panel, GuiFlowH("group_time"))
+  local group_time = GuiElement.add(right_panel, GuiFlowH("group_time"))
   GuiElement.add(group_time, GuiLabel("label_time"):caption({"helmod_data-panel.base-time", ""}):style("helmod_label_title_frame"))
   
   GuiElement.add(group_time, GuiDropDown(self.classname, "change-time", model_id):items(items, default_time))
@@ -467,25 +432,16 @@ end
 -- @param #LuaEvent event
 --
 function AbstractTab:updateIndexPanel(event)
-  local models = Model.getModels()
-  local model_id = User.getParameter("model_id")
-
   if self:hasIndexModel() then
-    -- index panel
-    local index_panel = self:getIndexPanel()
+    local models = Model.getModels()
+    local model_id = User.getParameter("model_id")
+  
+      -- index panel
+    local index_panel = self:getFramePanel("model_index")
     index_panel.clear()
+    local table_index = GuiElement.add(index_panel, GuiTable("table_index"):column(GuiElement.getIndexColumnNumber()):style(helmod_table_style.list))
     if Model.countModel() > 0 then
       local i = 0
-      -- local model_sorter = function(t,a,b) 
-      --   local elementB = Model.firstRecipe(t[b]["blocks"]) or {name="B"}
-      --   local elementA = Model.firstRecipe(t[a]["blocks"]) or {name="A"}
-      --   local order_by_name = elementB.name > elementA.name
-      --   if t[b]["group"] ~= nil and t[a]["group"] ~= nil then
-      --     return order_by_name and t[b]["group"] > t[a]["group"] 
-      --   else
-      --     return order_by_name and Model.countList(t[b]["blocks"]) < Model.countList(t[a]["blocks"]) 
-      --   end 
-      -- end
       for _,imodel in pairs(models) do
         i = i + 1
         local style = "helmod_button_default"
@@ -493,27 +449,26 @@ function AbstractTab:updateIndexPanel(event)
         if imodel.id == model_id then
           if element ~= nil then
             local tooltip = GuiTooltipModel("tooltip.info-model"):element(imodel)
-            GuiElement.add(index_panel, GuiButtonSprite(self.classname, "change-model", imodel.id):sprite(element.type, element.name):tooltip(tooltip))
+            GuiElement.add(table_index, GuiButtonSprite(self.classname, "change-model", imodel.id):sprite(element.type, element.name):tooltip(tooltip))
           else
-            local button = GuiElement.add(index_panel, GuiButton(self.classname, "change-model", imodel.id):sprite("menu", "help-white", "help"):style("helmod_button_menu_selected"))
+            local button = GuiElement.add(table_index, GuiButton(self.classname, "change-model", imodel.id):sprite("menu", "help-white", "help"):style("helmod_button_menu_selected"))
             button.style.width = 36
+            --button.style.height = 36
           end
         else
           if element ~= nil then
             local tooltip = GuiTooltipModel("tooltip.info-model"):element(imodel)
-            GuiElement.add(index_panel, GuiButtonSelectSprite(self.classname, "change-model", imodel.id):sprite(element.type, element.name):tooltip(tooltip))
+            GuiElement.add(table_index, GuiButtonSelectSprite(self.classname, "change-model", imodel.id):sprite(element.type, element.name):tooltip(tooltip))
           else
-            local button = GuiElement.add(index_panel, GuiButton(self.classname, "change-model", imodel.id):sprite("menu", "help-white", "help"):style("helmod_button_menu"))
+            local button = GuiElement.add(table_index, GuiButton(self.classname, "change-model", imodel.id):sprite("menu", "help-white", "help"):style("helmod_button_menu"))
             button.style.width = 36
+            --button.style.height = 36
           end
         end
 
       end
     end
-    GuiElement.add(index_panel, GuiButton(self.classname, "change-model", "new"):caption("+")).style.width=20
-  else
-    local index_panel = self:getIndexPanel()
-    index_panel.clear()
+    GuiElement.add(table_index, GuiButton(self.classname, "change-model", "new"):caption("+")).style.width=20
   end
 end
