+++ tab/AdminTab.lua
@@ -90,7 +90,7 @@ function AdminTab:getTab(panel_name, caption)
     return content_panel[scroll_name]
   end
   local tab_panel = GuiElement.add(content_panel, GuiTab(panel_name):caption(caption))
-  local scroll_panel = GuiElement.add(content_panel, GuiScroll(scroll_name):style(helmod_frame_style.scroll_pane):policy(true))
+  local scroll_panel = GuiElement.add(content_panel, GuiScroll(scroll_name):style("helmod_scroll_pane"):policy(true))
   content_panel.add_tab(tab_panel,scroll_panel)
   scroll_panel.style.horizontally_stretchable = true
   scroll_panel.style.vertically_stretchable = true
